+++ src/components/character-form/Tooltip.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../styles.module.css";
+import styles from "./styles.module.css";
 
 export default function Tooltip({ children }) {
   if (!children) {
